adj params and 64bit floats

+++ main.py
@@ -95,8 +95,8 @@ def anal(df, num_requested):
         # ####Model to fit prediction expression
         sess = tf.InteractiveSession()
 
-        x = tf.placeholder(tf.float64, shape=[None, num_factors])  # width of input
-        y_ = tf.placeholder(tf.float64, shape=[None, num_outs])  # width of output
+        x = tf.placeholder(tf.float32, shape=[None, num_factors])  # width of input
+        y_ = tf.placeholder(tf.float32, shape=[None, num_outs])  # width of output
 
         layer1 = tf.layers.dense(x, num_factors, tf.nn.softplus, bias_initializer=tf.random_uniform_initializer)
         layer2 = tf.layers.dense(layer1, num_factors, tf.nn.softplus, bias_initializer=tf.random_uniform_initializer)
